Generate patient

+++ src/com/company/Patient.java
@@ -8,12 +8,12 @@ public class Patient {
     private int day;
     private int month;
     private int year;
-    private int height;
+    private double height;
     private double weight;
     private String bloodType;
     private ArrayList<Record> records;
 
-    public Patient(String name, String surname, int day, int month, int year, int height, double weight, String bloodType) {
+    public Patient(String name, String surname, int day, int month, int year, double height, double weight, String bloodType) {
         this.name = name;
         this.surname = surname;
         this.day = day;
@@ -48,7 +48,7 @@ public class Patient {
         return year;
     }
 
-    public int getHeight() {
+    public double getHeight() {
         return height;
     }
